requirements.txt
torch==2.9.1
torchvision==0.24.1
numpy==2.4.2
pandas==2.3.2
opencv-python==4.12.0.88
Pillow==12.1.1
albumentations==2.0.8
PyYAML==6.0.3
tqdm==4.67.1
tifffile==2026.2.20
